+++ fujicv/data/__init__.py
@@ -2,6 +2,7 @@
 
 from fujicv.data.dataloader import build_dataloaders
 from fujicv.data.datasets import CSVImageDataset, build_splits
+from fujicv.data.hf_dataset import HFImageDataset, load_hf_dataset
 from fujicv.data.transforms import get_train_transforms, get_val_transforms
 
 __all__ = [
@@ -10,4 +11,6 @@ __all__ = [
     "build_dataloaders",
     "get_train_transforms",
     "get_val_transforms",
+    "HFImageDataset",
+    "load_hf_dataset",
 ]